Added TKTAuthGuestEmpty configuration option
Added TKTAuthGuestEmpty to allow fallback to native application
authorization on guest logins.  If set to on, then guests
will have a REMOTE_USER of "" (empty string).  Many http-auth
systems will then allow authentication to be handled by the next
highest authenticator (or a built-in default).

Updated doc and test suites as well.